+++ app/src/main/java/com/example/cattasticpos/domain/repository/InventoryRepository.kt
@@ -6,6 +6,6 @@ import kotlinx.coroutines.flow.Flow
 interface InventoryRepository {
     fun getAllInventory(): Flow<List<InventoryItem>>
     suspend fun insertInventoryItems(items: List<InventoryItem>)
-    suspend fun decrementStock(inventoryId: String, amount: Int)
-    suspend fun restockItem(itemId: String, addedAmount: Int)
+    suspend fun decrementStock(inventoryId: String, amount: Double)
+    suspend fun restockItem(itemId: String, addedAmount: Double)
 }
