+++ app/src/main/java/com/example/cattasticpos/domain/usecase/RestockItemUseCase.kt
@@ -5,7 +5,7 @@ import com.example.cattasticpos.domain.repository.InventoryRepository
 class RestockItemUseCase(
     private val inventoryRepository: InventoryRepository
 ) {
-    suspend operator fun invoke(itemId: String, addedAmount: Int) {
+    suspend operator fun invoke(itemId: String, addedAmount: Double) {
         if (addedAmount > 0) {
             inventoryRepository.restockItem(itemId, addedAmount)
         }
